+++ internal/rdb/inspect.go
@@ -21,7 +21,7 @@ func (r *RDB) AllQueues() ([]string, error) {
 	return r.client.SMembers(context.Background(), base.AllQueues).Result()
 }
 
-// KEYS[1] -> asynq:<qname>
+// KEYS[1] -> asynq:<qname>:pending
 // KEYS[2] -> asynq:<qname>:active
 // KEYS[3] -> asynq:<qname>:scheduled
 // KEYS[4] -> asynq:<qname>:retry
@@ -30,10 +30,13 @@ func (r *RDB) AllQueues() ([]string, error) {
 // KEYS[7] -> asynq:<qname>:processed:<yyyy-mm-dd>
 // KEYS[8] -> asynq:<qname>:failed:<yyyy-mm-dd>
 // KEYS[9] -> asynq:<qname>:paused
+//
+// ARGV[1] -> task key prefix
 var currentStatsCmd = redis.NewScript(`
 local res = {}
+local pendingTaskCount = redis.call("LLEN", KEYS[1])
 table.insert(res, KEYS[1])
-table.insert(res, redis.call("LLEN", KEYS[1]))
+table.insert(res, pendingTaskCount)
 table.insert(res, KEYS[2])
 table.insert(res, redis.call("LLEN", KEYS[2]))
 table.insert(res, KEYS[3])
@@ -60,6 +63,13 @@ table.insert(res, KEYS[8])
 table.insert(res, fcount)
 table.insert(res, KEYS[9])
 table.insert(res, redis.call("EXISTS", KEYS[9]))
+table.insert(res, "oldest_pending_since")
+if pendingTaskCount > 0 then
+	local id = redis.call("LRANGE", KEYS[1], -1, -1)[1]
+	table.insert(res, redis.call("HGET", ARGV[1] .. id, "pending_since"))
+else
+	table.insert(res, 0)
+end
 return res`)
 
 // CurrentStats returns a current state of the queues.
@@ -83,7 +93,7 @@ func (r *RDB) CurrentStats(qname string) (*base.Stats, error) {
 		base.ProcessedKey(qname, now),
 		base.FailedKey(qname, now),
 		base.PausedKey(qname),
-	}).Result()
+	}, base.TaskKeyPrefix(qname)).Result()
 	if err != nil {
 		return nil, errors.E(op, errors.Unknown, err)
 	}
@@ -127,6 +137,12 @@ func (r *RDB) CurrentStats(qname string) (*base.Stats, error) {
 				stats.Paused = false
 			} else {
 				stats.Paused = true
+			}
+		case "oldest_pending_since":
+			if val == 0 {
+				stats.Latency = 0
+			} else {
+				stats.Latency = r.clock.Now().Sub(time.Unix(0, int64(val)))
 			}
 		}
 	}
@@ -469,10 +485,6 @@ for _, id in ipairs(ids) do
 end
 return data
 `)
-
-// PENDING(GIL): TODO remove @merge
-// listMessages returns a list of TaskMessage in Redis list with the given key.
-//func (r *RDB) listMessages(key, qname string, pgn base.Pagination) ([]*base.TaskMessage, error) {
 
 // listMessages returns a list of TaskInfo in Redis list with the given key.
 func (r *RDB) listMessages(qname string, state base.TaskState, pgn base.Pagination) ([]*base.TaskInfo, error) {